requirements.txt
flask==3.0.2
openai==1.12.0
openai-whisper==20231117
python-dotenv==1.0.1
Werkzeug==3.0.1 